# Conflicts:
#	.gitmodules
#	src/Cool/Variables/Variables.h

+++ src/Cool/Log/internal/Message.h
@@ -14,7 +14,7 @@ struct Message {
     std::string body;
 
     std::chrono::system_clock::time_point timestamp;
-    uint64_t                              count;
+    uint32_t                              count;
 
     Message(Severity severity, std::string_view category, std::string_view body)
         : severity{severity}
